+++ galsim/phase_screens.py
@@ -756,15 +756,22 @@ class OpticalScreen(object):
         jmax = len(self.aberrations)-1
         maxn = _noll_to_zern(jmax)[0]
         shape = (maxn//2+1, maxn+1)  # (max power of |rho|^2,  max power of rho)
-        self.coef_array = np.zeros(shape, dtype=np.complex128)
+        self._coef_array = np.zeros(shape, dtype=np.complex128)
 
         noll_coef = _noll_coef_array(jmax, self.obscuration, self.annular_zernike)
-        self.coef_array = np.dot(noll_coef, self.aberrations[1:])
+        self._coef_array = np.dot(noll_coef, self.aberrations[1:])
         # Convert from unit disk coefficients to full aperture (diam != 2) coefficients.
-        self.coef_array /= (self.diam/2)**np.sum(np.mgrid[0:2*shape[0]:2, 0:shape[1]], axis=0)
+        self._coef_array /= (self.diam/2)**np.sum(np.mgrid[0:2*shape[0]:2, 0:shape[1]], axis=0)
 
         self.dynamic = False
         self.reversible = True
+
+    @property
+    def coef_array(self):
+        from .deprecated import depr
+        depr('optical_screen.coef_array', 1.5, 'optical_screen._coef_array',
+             'This is officially an implementation detail that users should not need to use.')
+        return self._coef_array
 
     def __str__(self):
         return "galsim.OpticalScreen(diam=%s, lam_0=%s)" % (self.diam, self.lam_0)
@@ -837,7 +844,7 @@ class OpticalScreen(object):
         # Note, this phase screen is actually independent of time and theta.
         r = u + 1j*v
         rsqr = np.abs(r)**2
-        return galsim.utilities.horner2d(rsqr, r, self.coef_array, dtype=complex).real * self.lam_0
+        return galsim.utilities.horner2d(rsqr, r, self._coef_array, dtype=complex).real * self.lam_0
 
     def wavefront_gradient(self, u, v, t=None, theta=None):
         """ Compute gradient of wavefront due to atmospheric phase screen.