+++ CHANGELOG.md
@@ -125,11 +125,12 @@ Deprecated Features
   obj.flux, gaussian.getSigma() -> gaussian.sigma, shear.getG1() -> shear.g1.
   There were many such methods, especially in the various kinds of GSObjects,
   as well as in Image, Bounds, Shear, and the Noise classes. (#904)
-- Deprecated ChromaticObject.obj -> _obj and is now an implementation detail,
-  not something that users should need to use. (#904)
+- Deprecated ChromaticObject.obj -> _obj, which is now an implementation
+  detail, not something that users should need to use. (#904)
 - Changed the objlist attribute of ChromaticSum and ChromaticConvolution to
   obj_list to conform to similar attributes in other classes. (#904)
-
+- Deprecated OpticalScreen.coef_array -> _coef_array, which is now an
+  implementation detail. (#904)
 
 
 New Features